+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [`validateApiRequest()`](./docs/forms.md#validateapirequest) function
+
 ## [0.11.2] - 2021-09-09
 
 ### Fixed