+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2021-09-15
+
 ### Added
 
 - [`validateApiRequest()`](./docs/forms.md#validateapirequest) function